moved events from main to a listener and remade the log output
