+++ resources/config.yml
@@ -1,7 +1,11 @@
 # Default Config File for PureEntitiesX
+---
+# when set to true the creeper will explode and also destroy blocks
 creeper:
-  block-breaking-explosion: false # when set to true the creeper will explode and also burst blocks
+  block-breaking-explosion: false
 
 # control logfile logging
+# possible: info = 0, warn = 1, debug = 2
 logfile:
-  loglevel: info # possible: debug, info, warn
+  loglevel: 0
+...
