+++ src/revivalpmmp/pureentities/entity/monster/walking/Skeleton.php
@@ -6,7 +6,6 @@ use revivalpmmp\pureentities\entity\monster\WalkingMonster;
 use pocketmine\entity\Entity;
 use pocketmine\entity\Projectile;
 use pocketmine\entity\ProjectileSource;
-use pocketmine\event\entity\EntityDamageByEntityEvent;
 use pocketmine\event\entity\EntityShootBowEvent;
 use pocketmine\event\entity\ProjectileLaunchEvent;
 use pocketmine\event\Timings;
