+++ src/revivalpmmp/pureentities/entity/monster/walking/SnowGolem.php
@@ -6,7 +6,6 @@ use revivalpmmp\pureentities\entity\monster\WalkingMonster;
 use pocketmine\entity\Entity;
 use pocketmine\entity\Projectile;
 use pocketmine\entity\ProjectileSource;
-use pocketmine\event\entity\EntityDamageByEntityEvent;
 use pocketmine\event\entity\ProjectileLaunchEvent;
 use pocketmine\item\Item;
 use pocketmine\level\sound\LaunchSound;
