+++ src/revivalpmmp/pureentities/task/AutoDespawnTask.php
@@ -17,11 +17,11 @@ class AutoDespawnTask extends PluginTask {
     }
     
     public function onRun($currentTick){
-        PureEntities::logDebug("AutoDespawnTask: onRun ($currentTick)");
+        PureEntities::logOutput("AutoDespawnTask: onRun ($currentTick)",PureEntities::DEBUG);
         $despawnable = [];
         foreach($this->plugin->getServer()->getLevels() as $level) {
             foreach($level->getEntities() as $entity) {
-                $despawnable[$entity->getId()] = 2; 
+                $despawnable[$entity->getId()] = 2;
                 foreach($level->getPlayers() as $player) {
                     if($player->distance($entity) <= 32) {
                         $despawnable[$entity->getId()] = 1;
@@ -34,14 +34,14 @@ class AutoDespawnTask extends PluginTask {
                     $probability = mt_rand(1, 100);
                     if($probability === 1) {
                         if($entity instanceof Monster) {
-                            PureEntities::logDebug("AutoDespawnTask: close monster entity (id: ". $entity->getId() . ", name:" . $entity->getNameTag() .")");
+                            PureEntities::logOutput("AutoDespawnTask: close monster entity (id: ". $entity->getId() . ", name:" . $entity->getNameTag() .")",PureEntities::DEBUG);
                             $entity->close();
                         }
                     }
                     
                 } elseif($despawnable[$entity->getId()] === 3) {
                     if($entity instanceof Animal || $entity instanceof Monster) {
-                        PureEntities::logDebug("AutoDespawnTask: close animal/monster entity (id: ". $entity->getId() . ", name:" . $entity->getNameTag() .")");
+                        PureEntities::logOutput("AutoDespawnTask: close animal/monster entity (id: ". $entity->getId() . ", name:" . $entity->getNameTag() .")",PureEntities::DEBUG);
                         $entity->close();
                     }
                 }
